+++ R/stats.R
@@ -171,7 +171,7 @@ mean_interval <- function(x, injury_event=FALSE) {
 #' draws are used. The event year and lag significance levels should match, at least in the general
 #' pattern.
 #'
-#' We note that our implementation of run_sea borrows from the \code{dplR:::sea} function in how it performs
+#' We note that our implementation of run_sea borrows from the \code{dplR::sea} function in how it performs
 #' the bootstrap procedure, but differs in the kind of output provided for the user.
 #'
 #' @return A list of three data frames, following the output of EVENT.
@@ -311,12 +311,12 @@ run_sea <- function(x, key, years_before=6, years_after=4,
 
   departure_table <- out_table[, -c(3, 4, 15, 16)]
   departure_table[, 2] <- key_event_table[, 2] - rand_event_table[, 2]
-  departure_table[, 3] <- apply(re.table, 2, function(x) -1 * 1.960*sd(x, na.rm=TRUE))
-  departure_table[, 4] <- apply(re.table, 2, function(x)      1.960*sd(x, na.rm=TRUE))
-  departure_table[, 5] <- apply(re.table, 2, function(x) -1 * 2.575*sd(x, na.rm=TRUE))
-  departure_table[, 6] <- apply(re.table, 2, function(x)      2.575*sd(x, na.rm=TRUE))
-  departure_table[, 7] <- apply(re.table, 2, function(x) -1 * 3.294*sd(x, na.rm=TRUE))
-  departure_table[, 8] <- apply(re.table, 2, function(x)      3.294*sd(x, na.rm=TRUE))
+  departure_table[, 3] <- apply(re.table, 2, function(x) -1 * 1.960*stats::sd(x, na.rm=TRUE))
+  departure_table[, 4] <- apply(re.table, 2, function(x)      1.960*stats::sd(x, na.rm=TRUE))
+  departure_table[, 5] <- apply(re.table, 2, function(x) -1 * 2.575*stats::sd(x, na.rm=TRUE))
+  departure_table[, 6] <- apply(re.table, 2, function(x)      2.575*stats::sd(x, na.rm=TRUE))
+  departure_table[, 7] <- apply(re.table, 2, function(x) -1 * 3.294*stats::sd(x, na.rm=TRUE))
+  departure_table[, 8] <- apply(re.table, 2, function(x)      3.294*stats::sd(x, na.rm=TRUE))
   temp <- apply(re.table, 2, function(x) stats::median(x))  # medians from simulated table
   departure_table[, 3] <- apply(re.table, 2, function(x) -1 * 1.960 * stats::sd(x, na.rm = TRUE))
   departure_table[, 4] <- apply(re.table, 2, function(x)      1.960 * stats::sd(x, na.rm = TRUE))
@@ -370,4 +370,77 @@ sample_depth <- function(a){
   out
 }
 
+#' Generate site-level summary statistics
+#'
+#' @param x An fhx object
+#' @param site_name Three character site code, defaults to "XXX"
+#' @param year_range Delimits the analysis period. For example, \code{c(1600, 1900)}.
+#' @param filter_prop An optional argument if the user chooses to include a composite rug in their plot. This is passed to \code{composite}. See this function for details.
+#' @param filter_min_rec An optional argument if the user chooses to include a composite rug in their plot. This is passed to \code{composite}. See this function for details.
+#' @param filter_min_scarred An optional argument if the user chooses to include a composite rug in their plot. This is passed to \code{composite}. See this function for details.
+#' @param injury_event Boolean indicating whether injuries should be considered recorders. This is passed to \code{composite}. See this function for details.
+#'
+#' @details This function prodiuces a summary table for any fhx object. The statistics it includes are shared by other popular fire history software such as FHX2 and FHAES.
+#' @return A data.frame of summary statistics
+#' @export
+
+site_stats <- function(x, site_name = 'XXX', year_range = NULL, filter_prop = 0.25, filter_min_rec = 2,
+                        filter_min_scarred = 1, injury_event = FALSE)
+{
+
+  stopifnot(is.fhx(x))
+  sumNames <- c('number_trees', 'first_year', 'last_year', 'first_event', 'last_event',
+                'number_intervals', 'mean_interval', 'median_interval',
+                'standard_dev', 'coef_var', 'min_interval', 'max_interval',
+                'weibull_shape', 'weibull_scale', 'weibull_mean',
+                'weibull_median', 'weibull_mode', 'KS_d', 'pval', 'lower_exceedance',
+                'upper_exceedance')
+  site.stats <- matrix(nrow=21, ncol=1, dimnames = list(sumNames, site_name))
+  # Perform site composite for interval stats
+  if (!is.null(year_range)) {
+    x <- x[x$year >= min(year_range) & x$year <= max(year_range), ]
+  }
+  x.comp <- composite(x, filter_prop = filter_prop, filter_min_rec = filter_min_rec,
+                          filter_min_scarred = filter_min_scarred, injury_event = injury_event)
+  intervals <- diff(get_event_years(x.comp)[[1]])
+  if(length(intervals) < 2)
+    stop("Too few fire intervals to compute a summary")
+  # Weibull fit
+  ft.r <- MASS::fitdistr(intervals, "weibull")
+  shape <- as.numeric(ft.r$estimate[1])
+  scale <- as.numeric(ft.r$estimate[2])
+  weib.quants <- stats::qweibull(c(.125, .5, .875), shape=shape, scale=scale)
+  gf <- suppressWarnings( stats::ks.test(intervals, y=stats::pweibull, shape=shape, scale=scale, alternative='less'))
+  # Fill out summary table
+  site.stats['number_trees', ] <- length(levels(x$series))
+  site.stats['first_year', ] <- first_year(x)
+  site.stats['last_year', ] <- last_year(x)
+  if (injury_event == FALSE) {
+    site.stats['first_event', ] <- min(x[grep('fs', x$rec_type), ]$year)
+    site.stats['last_event', ] <- max(x[grep('fs', x$rec_type), ]$year)
+  }
+  else {
+    site.stats['first_event', ] <- min(min(x[grep('fs', x$rec_type), ]$year),
+                                       min(x[grep('fi', x$rec_type), ]$year))
+    site.stats['last_event', ] <- max(max(x[grep('fs', x$rec_type), ]$year),
+                                      max(x[grep('fi', x$rec_type), ]$year))
+  }
+  site.stats['number_intervals', ] <- length(intervals)
+  site.stats['mean_interval', ] <- round(mean(intervals), 1)
+  site.stats['median_interval', ] <- round(stats::median(intervals), 1)
+  site.stats['standard_dev', ] <- round(stats::sd(intervals), 2)
+  site.stats['coef_var', ] <- round(stats::sd(intervals)/mean(intervals), 2)
+  site.stats['min_interval', ] <- min(intervals)
+  site.stats['max_interval', ] <- max(intervals)
+  site.stats['weibull_shape', ] <- round(shape, 2)
+  site.stats['weibull_scale', ] <- round(scale, 2)
+  site.stats['weibull_mean', ] <- round(scale * gamma(1 + 1/shape), 2)
+  site.stats['weibull_median', ] <- round(weib.quants[2], 2)
+  site.stats['weibull_mode', ] <- round(scale * ((shape-1)/shape)^(1/shape), 2)
+  site.stats['KS_d', ] <- round(gf$statistic, 2)
+  site.stats['pval', ] <- round(gf$p.value, 2)
+  site.stats['lower_exceedance', ] <- round(weib.quants[1], 2)
+  site.stats['upper_exceedance', ] <- round(weib.quants[3], 2)
+  return(site.stats)
+}